Reorganize top-level headers in libnetdata.
Move top-level libnetdata headers (simple hashtable,
bitmap, linked list, xxHash) into more specific
subdirectories and update the corresponding
include paths throughout the project to reflect
these changes.
