+++ src/collectors/windows-events.plugin/windows-events-providers.c
@@ -64,7 +64,7 @@ typedef struct provider {
 #define SIMPLE_HASHTABLE_VALUE2KEY_FUNCTION provider_value_to_key
 #define SIMPLE_HASHTABLE_COMPARE_KEYS_FUNCTION provider_cache_compar
 #define SIMPLE_HASHTABLE_SAMPLE_IMPLEMENTATION 1
-#include "libnetdata/simple_hashtable.h"
+#include "libnetdata/simple_hashtable/simple_hashtable.h"
 
 static struct {
     SPINLOCK spinlock;
