+++ src/libnetdata/common.h
@@ -252,7 +252,7 @@ typedef uint32_t uid_t;
 #endif
 
 #define XXH_INLINE_ALL
-#include "xxhash.h"
+#include "xxHash/xxhash.h"
 
 // --------------------------------------------------------------------------------------------------------------------
 // OpenSSL
@@ -378,7 +378,7 @@ typedef uint32_t uid_t;
 // --------------------------------------------------------------------------------------------------------------------
 // Macro-only includes
 
-#include "linked-lists.h"
+#include "linked_lists/linked_lists.h"
 
 // --------------------------------------------------------------------------------------------------------------------
