+++ src/libnetdata/config/dyncfg.c
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-#include "../../libnetdata/libnetdata.h"
+#include "../libnetdata.h"
 
 // ----------------------------------------------------------------------------
