+++ src/libnetdata/linked_lists/linked_lists.h
