+++ src/libnetdata/local-sockets/local-sockets.h
@@ -29,7 +29,7 @@
 
 #define SIMPLE_HASHTABLE_VALUE_TYPE uint64_t
 #define SIMPLE_HASHTABLE_NAME _NET_NS
-#include "libnetdata/simple_hashtable.h"
+#include "libnetdata/simple_hashtable/simple_hashtable.h"
 
 // --------------------------------------------------------------------------------------------------------------------
 // hashtable for keeping the sockets of PIDs
@@ -38,7 +38,7 @@
 struct pid_socket;
 #define SIMPLE_HASHTABLE_VALUE_TYPE struct pid_socket
 #define SIMPLE_HASHTABLE_NAME _PID_SOCKET
-#include "libnetdata/simple_hashtable.h"
+#include "libnetdata/simple_hashtable/simple_hashtable.h"
 
 // --------------------------------------------------------------------------------------------------------------------
 // hashtable for keeping all the sockets
@@ -47,7 +47,7 @@ struct pid_socket;
 struct local_socket;
 #define SIMPLE_HASHTABLE_VALUE_TYPE struct local_socket
 #define SIMPLE_HASHTABLE_NAME _LOCAL_SOCKET
-#include "libnetdata/simple_hashtable.h"
+#include "libnetdata/simple_hashtable/simple_hashtable.h"
 
 // --------------------------------------------------------------------------------------------------------------------
 // hashtable for keeping all local IPs
@@ -56,7 +56,7 @@ struct local_socket;
 union ipv46;
 #define SIMPLE_HASHTABLE_VALUE_TYPE union ipv46
 #define SIMPLE_HASHTABLE_NAME _LOCAL_IP
-#include "libnetdata/simple_hashtable.h"
+#include "libnetdata/simple_hashtable/simple_hashtable.h"
 
 // --------------------------------------------------------------------------------------------------------------------
 // hashtable for keeping all listening ports
@@ -65,7 +65,7 @@ union ipv46;
 struct local_port;
 #define SIMPLE_HASHTABLE_VALUE_TYPE struct local_port
 #define SIMPLE_HASHTABLE_NAME _LISTENING_PORT
-#include "libnetdata/simple_hashtable.h"
+#include "libnetdata/simple_hashtable/simple_hashtable.h"
 
 // --------------------------------------------------------------------------------------------------------------------
