+++ src/libnetdata/os/system-maps/cached-sid-username.c
@@ -31,7 +31,7 @@ typedef struct {
 #define SIMPLE_HASHTABLE_VALUE2KEY_FUNCTION sid_value_to_key
 #define SIMPLE_HASHTABLE_COMPARE_KEYS_FUNCTION sid_cache_compar
 #define SIMPLE_HASHTABLE_SAMPLE_IMPLEMENTATION 1
-#include "libnetdata/simple_hashtable.h"
+#include "libnetdata/simple_hashtable/simple_hashtable.h"
 
 static struct {
     SPINLOCK spinlock;
