+++ src/libnetdata/xxHash/xxhash.h
